requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-engineio==4.9.1
paho-mqtt==1.6.1
